tools/requirements.txt
Pillow>=9.0.0
tomli>=2.0.0; python_version < "3.11"
